# test_pvc_creation_deletion_performance.py - fixing deletion measurements