--- a/tests/e2e/performance/csi_tests/test_pvc_creation_deletion_performance.py
+++ b/tests/e2e/performance/csi_tests/test_pvc_creation_deletion_performance.py
@@ -9,6 +9,7 @@ import ocs_ci.ocs.exceptions as ex
 import statistics
 import tempfile
 import yaml
+import time
 
 from ocs_ci.framework.testlib import performance
 from ocs_ci.ocs.perftests import PASTest
@@ -308,7 +309,8 @@ class TestPVCCreationDeletionPerformance(PASTest):
             f'get sc {Interface_Info[self.interface]["sc"]} -o jsonpath="'
             + '{.reclaimPolicy}"'
         )
-        if rec_policy == constants.RECLAIM_POLICY_DELETE:
+
+        if rec_policy[0].strip('"') == constants.RECLAIM_POLICY_DELETE:
             log.info("Wait for all PVC(s) backed PV(s) to be deleted")
             # Timeout for each PV to be deleted is 20 sec.
             performance_lib.wait_for_resource_bulk_status(