Previous version of the test was measuring Deletion and CSI Deletion times as None. 
Following the current fix Deletion and CSI deletion times will be measured. 

Example from test log: 

17:44:58 - MainThread - test_pvc_creation_deletion_performance - INFO  - Interface: CephBlockPool, PVC size: 5Gi. PVC number 1 was created in 0.087 seconds.
17:44:58 - MainThread - test_pvc_creation_deletion_performance - INFO  - Interface: CephBlockPool, PVC size: 5Gi. PVC number 1 was deleted in 0.078 seconds.
17:44:58 - MainThread - test_pvc_creation_deletion_performance - INFO  - Interface: CephBlockPool, PVC size: 5Gi. PVC number 1 was csi_created in 0.082 seconds.
17:44:58 - MainThread - test_pvc_creation_deletion_performance - INFO  - Interface: CephBlockPool, PVC size: 5Gi. PVC number 1 was csi_deleted in 0.066 seconds.
17:44:58 - MainThread - test_pvc_creation_deletion_performance - INFO  - Interface: CephBlockPool, PVC size: 5Gi. PVC number 2 was created in 0.079 seconds.
17:44:58 - MainThread - test_pvc_creation_deletion_performance - INFO  - Interface: CephBlockPool, PVC size: 5Gi. PVC number 2 was deleted in 0.08 seconds.
17:44:58 - MainThread - test_pvc_creation_deletion_performance - INFO  - Interface: CephBlockPool, PVC size: 5Gi. PVC number 2 was csi_created in 0.074 seconds.
17:44:58 - MainThread - test_pvc_creation_deletion_performance - INFO  - Interface: CephBlockPool, PVC size: 5Gi. PVC number 2 was csi_deleted in 0.07 seconds.
17:44:58 - MainThread - test_pvc_creation_deletion_performance - INFO  - Interface: CephBlockPool, PVC size: 5Gi. PVC number 3 was created in 0.081 seconds.
17:44:58 - MainThread - test_pvc_creation_deletion_performance - INFO  - Interface: CephBlockPool, PVC size: 5Gi. PVC number 3 was deleted in 0.078 seconds.
17:44:58 - MainThread - test_pvc_creation_deletion_performance - INFO  - Interface: CephBlockPool, PVC size: 5Gi. PVC number 3 was csi_created in 0.075 seconds.
17:44:58 - MainThread - test_pvc_creation_deletion_performance - INFO  - Interface: CephBlockPool, PVC size: 5Gi. PVC number 3 was csi_deleted in 0.069 seconds.
17:44:58 - MainThread - test_pvc_creation_deletion_performance - INFO  - Interface: CephBlockPool, PVC size: 5Gi. PVC number 4 was created in 0.08 seconds.
17:44:58 - MainThread - test_pvc_creation_deletion_performance - INFO  - Interface: CephBlockPool, PVC size: 5Gi. PVC number 4 was deleted in 0.079 seconds.
17:44:58 - MainThread - test_pvc_creation_deletion_performance - INFO  - Interface: CephBlockPool, PVC size: 5Gi. PVC number 4 was csi_created in 0.072 seconds.
17:44:58 - MainThread - test_pvc_creation_deletion_performance - INFO  - Interface: CephBlockPool, PVC size: 5Gi. PVC number 4 was csi_deleted in 0.065 seconds.
17:44:58 - MainThread - test_pvc_creation_deletion_performance - INFO  - Interface: CephBlockPool, PVC size: 5Gi. PVC number 5 was created in 0.081 seconds.
17:44:58 - MainThread - test_pvc_creation_deletion_performance - INFO  - Interface: CephBlockPool, PVC size: 5Gi. PVC number 5 was deleted in 0.084 seconds.
17:44:58 - MainThread - test_pvc_creation_deletion_performance - INFO  - Interface: CephBlockPool, PVC size: 5Gi. PVC number 5 was csi_created in 0.077 seconds.
17:44:58 - MainThread - test_pvc_creation_deletion_performance - INFO  - Interface: CephBlockPool, PVC size: 5Gi. PVC number 5 was csi_deleted in 0.073 seconds.

Signed-off-by: Yulia Persky <ypersky@redhat.com>